Apply linksys qsdk patch : 050_ipv6_defrag_timeout.patch

+++ net/ipv6/netfilter/nf_conntrack_reasm.c
@@ -1,3 +1,4 @@
+/* Modifications were made by Linksys on or before Fri Dec  2 14:44:09 PST 2016 */
 /*
  * IPv6 fragment reassembly for connection tracking
  *
@@ -322,7 +323,7 @@ found:
 
 	/* Check for overlap with preceding fragment. */
 	if (prev &&
-	    (NFCT_FRAG6_CB(prev)->offset + prev->len) > offset)
+	    (NFCT_FRAG6_CB(prev)->offset + (int)prev->len) > offset)
 		goto discard_fq;
 
 	/* Look for overlap with succeeding segment. */
